Fixed issue with switching php versions
I had an issue with switching phpversions using valet use ...
When using brew unlink/link manually all went well. Bu when using valet use ... the switching went well, but the restart of the phpfpm services caused an error. This is due to the fact that valet uses the 'sudo' command for restarting the services. The services then get restarted as the 'root' user instead of the current user.

I changed the command to 'quietlyAsUser' to restart the services. This fixed the problem.