php version switch bug in restarting php services
Valet uses the sudo command to restart the php services after switching a php version. This caused the valet use command to be not switch my php version.
By chaning the restart commando to use 'sudo -u user()' the current user is used to restart the brew services. This is the same when restarting brew services manually (which worked flawlessly).